+++ test/test325.js
@@ -22,7 +22,8 @@ describe('Test 325 IDENTITY', function() {
       msgid  INT          NOT NULL IDENTITY ,
       msgts  DATETIME     NOT NULL DEFAULT(CURRENT_TIMESTAMP),
       msg    VARCHAR(MAX) NOT NULL,
-      status VARCHAR(20)  NOT NULL DEFAULT('new'),
+      status VARCHAR(20)  NOT NULL DEFAULT('new') 
+        CHECK(status IN('new', 'open')),
       CONSTRAINT PK_Messages 
         PRIMARY KEY NONCLUSTERED(msgid),
       CONSTRAINT UNQ_Messages_status_msgid 
@@ -69,7 +70,7 @@ describe('Test 325 IDENTITY', function() {
     done();
   });
 
-    it('9. INSERT INTO with IDENTITY',function(done){
+    it('7. INSERT INTO with IDENTITY',function(done){
       // console.log(69,alasql.tables.Messages.identities);
       // console.log(69,alasql.tables.Messages.uniqs);
       // console.log(69,alasql.tables.Messages.pk);
@@ -80,29 +81,36 @@ describe('Test 325 IDENTITY', function() {
     });
 
 
-  it('7. INSERT INTO with IDENTITY',function(done){
+  it('8. INSERT INTO with IDENTITY',function(done){
     var res = alasql('INSERT INTO dbo.Messages (msg, status) \
       VALUES("I hate you!","new")');
     assert(res == 1);
     done();
   });
 
-  it('8. INSERT INTO with IDENTITY',function(done){
+  it('9. INSERT INTO with IDENTITY',function(done){
     var res = alasql('INSERT INTO dbo.Messages (msg, status) \
       VALUES("I hate you to much!","new")');
     assert(res == 1);
     done();
   });
 
-  it('9. INSERT INTO with IDENTITY',function(done){
+  it('10. INSERT INTO with IDENTITY',function(done){
     var res = alasql('SELECT COLUMN msgid FROM dbo.Messages');
     assert.deepEqual(res,[1,2,3]);
 //    console.log(res);
     done();
   });
 
+  it('11. CHECK CONSTRAINT on column',function(done){
+    assert.throws(function(){
+      var res = alasql('INSERT INTO dbo.Messages (msg, status) \
+        VALUES("It is not so bad","done")');
+    },Error);
+    done();
+  });
 
-  it('3. DROP DATABASE',function(done){
+  it('99. DROP DATABASE',function(done){
     alasql('DROP DATABASE test325');
     done();
   });